requirements.txt
anthropic>=0.40.0
python-telegram-bot>=21.0
pytz>=2024.1
